test(dedupe): consolidate trading_stats tests (5 → 3 files)
Consolidate TradingStatsService test files in tests/unit/gpt_trader/tui/services:
- test_trading_stats_service_model.py + singleton.py + windows.py → test_trading_stats.py
- test_trading_stats_service_compute_stats.py → test_trading_stats_compute.py (rename)
- test_trading_stats_service_trade_matching.py → test_trading_stats_matching.py (rename)

Cluster: 5b3d06876e1e
Files: 5 → 3 (-2)
Tests: 19 passing
