+++ tests/_triage/dedupe_candidates.yaml
@@ -2416,7 +2416,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/312
     added: '2026-01-20'
   68f5edc7f33a:
     type: similar_names
