+++ tests/unit/gpt_trader/tui/services/test_trading_stats_compute.py
